backend/requirements.txt
pymongo
dnspython
flask
pyjwt
Pillow
matplotlib